requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.3
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0 